+++ mailosaur/analysis_model.go
@@ -1,4 +1,4 @@
-package mailosaur_go
+package mailosaur
 
 type SpamAnalysisResult struct {
 	SpamFilterResults SpamFilterResults
