+++ mailosaur/analysis_operations.go
@@ -1,4 +1,4 @@
-package mailosaur_go
+package mailosaur
 
 type AnalysisOperations struct {
 	client *MailosaurClient
