+++ tests/resources/cat.png
