+++ tests/resources/dog.png
